refactor errors
