+++ ngmix/prepsfadmom/__init__.py
@@ -4,3 +4,5 @@ from . import prepsfadmom
 from .prepsfadmom import *
 
 from . import prepsfadmom_nb
+from . import models
+from . import errors
